Disable inline keyed-access fast path, fix clippy warnings
The inline store/load keyed fast paths have unresolved correctness
issues (SIGSEGV on sieve_primes_1k, array_push_sum_1k).  Bypass
the IC fast paths entirely: all keyed loads/stores go through the
safe runtime stubs.  The function remains JIT-compiled for all
other operations (arithmetic, named property access, calls, etc.).

Also fixes two clippy warnings:
- Unused variable vec_layout in emit_inline_store_keyed_smi
- Unnecessary cast (k as i64) in jit_runtime_fusion_resolve

+++ crates/stator_jse/src/compiler/baseline/compiler.rs
@@ -3315,7 +3315,7 @@ pub(crate) mod jit_runtime {
         // fusion calls from Maglev can skip the runtime resolve.
         FUSION_CTX_CALLEE.store(callee_i64, Ordering::Relaxed);
         FUSION_CTX_SLOT_PTR.store(slot_ptr, Ordering::Relaxed);
-        FUSION_CTX_K.store(k as i64, Ordering::Relaxed);
+        FUSION_CTX_K.store(k, Ordering::Relaxed);
 
         FusionResolved { slot_ptr, k }
     }
